requirements.txt
streamlit>=1.22.0
pillow>=9.3.0        # Pour PIL
pandas>=1.5.0
plotly>=5.11.0
yfinance>=0.2.18
numpy>=1.23.0
fredapi>=0.5.0
pymysql>=1.0.2
scipy>=1.9.0
matplotlib>=3.6.0
tqdm>=4.64.1
python-dotenv>=0.21.0